Updated dev file for avatargroup
